move keyedMap to utils
